requirements.txt
google-generativeai==0.3.0
streamlit==1.28.0
streamlit-chat==0.1.0
requests==2.31.0
beautifulsoup4==4.12.2
python-dateutil==2.8.2
python-dotenv==1.0.0
wolframalpha==5.0.0
